Compute, propage and use row/uniqueness preservation for joins between base tables
Previously, when a foreign key join referenced a dervied table, all base tables
within such derived table were assumed to always preserve row/uniqueness,
regardless of the joins they were involved in.  This caused foreign key joins
to be accepted that should have been rejected.  Such an example has been
added to the test suite, and here is what happened before the fix for that case:

     JOIN orders FOR KEY (id) <- order_items (order_id)
     RIGHT JOIN shipments FOR KEY (id) <- orders (shipment_id)
 ) AS q JOIN packages FOR KEY (shipment_id) -> q (shipment_id);
-ERROR:  foreign key join violation
-LINE 15: ) AS q JOIN packages FOR KEY (shipment_id) -> q (shipment_id...
-                              ^
-DETAIL:  referenced relation does not preserve uniqueness of keys
+ package_id | order_id | shipment_id | order_item_id
+------------+----------+-------------+---------------
+        900 |       10 |         100 |          1001
+        900 |       10 |         100 |          1000
+(2 rows)

The thinko was that while it's true we don't need to consider the effect
on row/uniqueness preservation, for base table instances that are referenced
by foreign key joins at the same query level, in that case they always
preserve row/uniqueness, but when used by derived tables from outer
query levels, we need to compute and use preservation.

Another change is that RLS is now ignored, since it's not part of
the SQL-standard, and because the logic would be more complicated.
Since RLS *could* be supported properly, a TODO item has been added
to the test that explains how it should work.